feat: refactor modal manager to sequence controller
- Moved ModalManagerController logic to SequenceController
- Updated endpoint for config sequence to include '/modal'
- Removed unused imports and methods related to explanations
